Namespace reorganization
