+++ CliToolkit.Tests/Templates/ValidApp.cs
@@ -1,6 +1,6 @@
 using System;
 using CliToolkit.Arguments;
-using CliToolkit.Arguments.Styles;
+using CliToolkit.Styles;
 
 namespace CliToolkit.Tests.Templates
 {
